+++ bittytax/conv/bittytax_conv.py
@@ -43,7 +43,7 @@ def main():
     parser.add_argument('-d',
                         '--debug',
                         action='store_true',
-                        help="enabled debug logging")
+                        help="enable debug logging")
     parser.add_argument('-uc',
                         '--unconfirmed',
                         action='store_true',